@@ -14,8 +14,8 @@
     "babel-loader": "^5.4.2",
     "http-server": "^0.9.0",
     "re-base": "^3.0.0",
-    "react": "^15.3.0",
-    "react-dom": "^15.3.0",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
     "webpack": "^1.13.1"
   }
 }
