@@ -20,11 +20,11 @@
   "dependencies": {
     "axios": "^0.13.1",
     "firebase": "^3.3.0",
+    "prop-types": "^15.6.0",
     "re-base": "^3.0.0",
-    "react": "^15.3.0",
-    "react-dom": "^15.3.0",
-    "react-router": "^2.6.1",
-    "reactfire": "^1.0.0"
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
+    "react-router-dom": "^4.2.2"
   },
   "devDependencies": {
     "babel-core": "^5.8.38",
