@@ -17,7 +17,7 @@
   },
   "dependencies": {
     "re-base": "^3.0.0",
-    "react": "^15.3.0",
-    "react-dom": "^15.3.0"
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0"
   }
 }
