@@ -1,15 +1,16 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { withRouter } from 'react-router';
+import PropTypes from 'prop-types';
+import { withRouter } from 'react-router-dom';
 
-class SearchGithub extends React.Component{
-  handleSubmit(e){
+class SearchGithub extends React.Component {
+  handleSubmit(e) {
     e.preventDefault();
     var username = ReactDOM.findDOMNode(this.refs.username).value;
     ReactDOM.findDOMNode(this.refs.username).value = '';
-    this.props.router.push(`/profile/${username}`);
+    this.props.history.push(`/profile/${username}`);
   }
-  render(){
+  render() {
     return (
       <div className="col-sm-12">
         <form onSubmit={this.handleSubmit.bind(this)}>
@@ -17,16 +18,18 @@ class SearchGithub extends React.Component{
             <input type="text" className="form-control" ref="username" />
           </div>
           <div className="form-group col-sm-5">
-            <button type="submit" className="btn btn-block btn-primary">Search Github </button>
+            <button type="submit" className="btn btn-block btn-primary">
+              Search Github
+            </button>
           </div>
         </form>
       </div>
-    )
+    );
   }
-};
+}
 
 SearchGithub.contextTypes = {
-  router: React.PropTypes.object.isRequired
+  router: PropTypes.object.isRequired
 };
 
 export default withRouter(SearchGithub);
