@@ -2,13 +2,13 @@ import React from 'react';
 import Main from '../components/Main';
 import Home from '../components/Home';
 import Profile from '../components/Profile';
-import { Router, Route, IndexRoute, browserHistory } from 'react-router';
+import { BrowserRouter as Router, Route } from 'react-router-dom';
 
 export default (
-    <Router history={browserHistory}>
-      <Route path="/" component={Main}>
-        <Route path="profile/:username" component={Profile} />
-        <IndexRoute component={Home} />
-      </Route>
-    </Router>
+  <Router>
+    <Main>
+      <Route exact path="/" component={Home} />
+      <Route exact path="/profile/:username" component={Profile} />
+    </Main>
+  </Router>
 );
