@@ -1,13 +1,13 @@
 module.exports = {
-  entry: "./app/App.js",
+  entry: './app/App.js',
   output: {
-    filename: "public/bundle.js"
+    filename: 'public/bundle.js'
   },
   module: {
     loaders: [
       {
         test: /\.jsx?$/,
-        exclude: /(node_modules|bower_components)/,
+        exclude: /node_module/,
         loader: 'babel-loader'
       }
     ]
